Leave asserts unchanged
Your code is also right, but the idea is to leave the asserts unchanged (so, it just should be coded the other way around ;-) )
In the 'newer' version, the var assert = require("assert") is build-in. But not very interesting to update that for all the other katas I think (is self-explanatory).